requirements.txt
flask
flask-cors
transformers
numpy
matplotlib
pillow
torchvision==0.18.0+cu121
torch==2.3.0